extract skeleton_builders from ConditionExampleGenerator
